vérification connecter pour profil

+++ src/Controller/ProfilController.php
@@ -74,15 +74,21 @@ class ProfilController extends AbstractController
     /**
      * @Route("/profil/delete/{id}", name="parents_delete", methods={"DELETE"})
      */
-    public function delete(Request $request, Parents $parent, TokenStorageInterface $tokenStorage): Response
+    public function delete(Request $request, Parents $parent, TokenStorageInterface $tokenStorage, AuthorizationCheckerInterface $authChecker): Response
     {
-        if ($this->isCsrfTokenValid('delete' . $parent->getId(), $request->request->get('_token'))) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $this->get('security.token_storage')->setToken(null);
-            $entityManager->remove($parent);
-            $entityManager->flush();
-        }
+        if (true === $authChecker->isGranted('ROLE_PARENT')) {
+            if ($this->isCsrfTokenValid('delete' . $parent->getId(), $request->request->get('_token'))) {
+                $entityManager = $this->getDoctrine()->getManager();
+                $this->get('security.token_storage')->setToken(null);
+                $entityManager->remove($parent);
+                $entityManager->flush();
 
-        return $this->redirectToRoute('accueil');
+                return $this->redirectToRoute('accueil');
+            }
+        } else {
+            return $this->render('403/403.html.twig', [
+                'erreur' => 'ACCES FORBIDEN'
+            ]);
+        }
     }
 }
